requirements.txt
streamlit==1.35.0
requests==2.31.0
langchain-core==0.1.52
python-dotenv==1.0.1